+++ Photo-Editor/Photo-Editor/Views/EditImageView.swift
@@ -7,11 +7,10 @@
 
 import SwiftUI
 import Photos
-import CoreImage
-import CoreImage.CIFilterBuiltins
 
 struct EditImageView: View {
     @State var image: UIImage
+    @State private var filterController = EditImageController()
     var onSave: (Any) -> Void
     @Binding var isLoading: Bool
     @State private var uploadStatusMessage: String?
@@ -31,7 +30,6 @@ struct EditImageView: View {
         _isLoading = isLoading
         self.onSave = onSave
     }
-
 
     var body: some View {
         VStack {
@@ -66,7 +64,6 @@ struct EditImageView: View {
             }
             .padding()
 
-
             HStack {
                 Button(action: uploadImageToCloudinary) {
                     Text("Upload and Save")
@@ -95,7 +92,6 @@ struct EditImageView: View {
                 dismiss()
             }
         }
-
         .sheet(isPresented: $showFilterDrawer) {
             FilterSelectionView(currentImage: originalImage, selectedFilter: $selectedFilter) { filter in
                 applyFilterAndRotation(filter)
@@ -105,38 +101,22 @@ struct EditImageView: View {
         }
     }
 
+    // MARK: - Rotation Actions
+
     func rotateImageLeft() {
-        rotationAngle -= 90
-        rotationAngle = rotationAngle.truncatingRemainder(dividingBy: 360)
-        applyRotation()
+        image = filterController.rotateImageLeft(currentImage: currentFilteredImage, currentAngle: &rotationAngle)
     }
 
     func rotateImageRight() {
-        rotationAngle += 90
-        rotationAngle = rotationAngle.truncatingRemainder(dividingBy: 360)
-        applyRotation()
+        image = filterController.rotateImageRight(currentImage: currentFilteredImage, currentAngle: &rotationAngle)
     }
 
+    // MARK: - Filter and Rotation
 
-
-
-
-
-    func rotateImage(by degrees: CGFloat) {
-        let radians = degrees * (.pi / 180)
-        let newSize = CGSize(width: image.size.height, height: image.size.width)
-        let renderer = UIGraphicsImageRenderer(size: newSize)
-
-        let rotatedImage = renderer.image { context in
-            context.cgContext.translateBy(x: newSize.width / 2, y: newSize.height / 2)
-            context.cgContext.rotate(by: radians)
-            context.cgContext.translateBy(x: -image.size.width / 2, y: -image.size.height / 2)
-            image.draw(at: .zero)
-        }
-
-        image = rotatedImage
+    func applyFilterAndRotation(_ filter: FilterType) {
+        image = filterController.applyFilterAndRotation(to: originalImage, filter: filter, rotationAngle: rotationAngle)
+        currentFilteredImage = image
     }
-
     
     private func uploadImageToCloudinary() {
         isLoading = true
@@ -167,125 +147,6 @@ struct EditImageView: View {
             } else {
                 showErrorAlert = true
             }
-        }
-    }
-
-    func applyFilterAndRotation(_ filter: FilterType) {
-        switch filter {
-        case .original:
-            currentFilteredImage = originalImage
-        case .cold:
-            currentFilteredImage = applyTemperatureFilter(to: originalImage, temperature: 4500)
-        case .warm:
-            currentFilteredImage = applyTemperatureFilter(to: originalImage, temperature: 8500)
-        }
-        applyRotation()
-    }
-
-
-
-    
-    func applyTemperatureFilter(to inputImage: UIImage, temperature: CGFloat) -> UIImage {
-        let ciContext = CIContext()
-        let ciImage = CIImage(image: inputImage)!
-        let filter = CIFilter.temperatureAndTint()
-        filter.inputImage = ciImage
-        filter.neutral = CIVector(x: temperature, y: 0)
-
-        if let outputImage = filter.outputImage,
-           let cgImage = ciContext.createCGImage(outputImage, from: outputImage.extent) {
-            return UIImage(cgImage: cgImage)
-        }
-
-        return inputImage
-    }
-    
-    func applyRotation() {
-        let radians = rotationAngle * (.pi / 180)
-        let newSize = CGSize(
-            width: abs(cos(radians)) * currentFilteredImage.size.width + abs(sin(radians)) * currentFilteredImage.size.height,
-            height: abs(sin(radians)) * currentFilteredImage.size.width + abs(cos(radians)) * currentFilteredImage.size.height
-        )
-
-        let renderer = UIGraphicsImageRenderer(size: newSize)
-
-        let rotatedImage = renderer.image { context in
-            context.cgContext.translateBy(x: newSize.width / 2, y: newSize.height / 2)
-            context.cgContext.rotate(by: radians)
-            context.cgContext.translateBy(x: -currentFilteredImage.size.width / 2, y: -currentFilteredImage.size.height / 2)
-            currentFilteredImage.draw(at: .zero)
-        }
-
-        image = rotatedImage
-    }
-
-
-
-
-}
-
-// MARK: - FilterType Enum
-
-enum FilterType {
-    case original, cold, warm
-}
-
-
-struct FilterSelectionView: View {
-    let currentImage: UIImage
-    @Binding var selectedFilter: FilterType
-    var onSelect: (FilterType) -> Void
-
-    var body: some View {
-        VStack {
-            Text("Select a Filter")
-                .font(.headline)
-                .padding()
-
-            HStack(spacing: 20) {
-                filterButton(for: .original, label: "Original", color: Color.gray)
-                filterButton(for: .cold, label: "Cold", color: Color.blue)
-                filterButton(for: .warm, label: "Warm", color: Color.orange)
-            }
-            .padding()
-        }
-    }
-
-    func filterButton(for filter: FilterType, label: String, color: Color) -> some View {
-        VStack {
-            color
-                .frame(width: selectedFilter == filter ? 100 : 80, height: selectedFilter == filter ? 100 : 80)
-                .cornerRadius(8)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 8)
-                        .stroke(selectedFilter == filter ? Color.white : Color.clear, lineWidth: 3)
-                )
-            Text(label)
-                .font(.caption)
-                .foregroundColor(.white)
-        }
-        .onTapGesture {
-            selectedFilter = filter
-            onSelect(filter)
-        }
-    }
-}
-
-
-
-
-struct FilterPreview: View {
-    let image: UIImage
-    let label: String
-    let applyFilter: (UIImage) -> UIImage
-
-    var body: some View {
-        VStack {
-            Image(uiImage: applyFilter(image))
-                .resizable()
-                .frame(width: 80, height: 80)
-                .cornerRadius(8)
-            Text(label)
         }
     }
 }
